Adds `db.loc` parameter to FacileDasetSet constructor to enable use of alt db for meta info
I added this because I first thought banging on the same db from muliple processes (ie. batchjobs) was messing around with functionality. Even though the database is read-only, temp tables are created when data is copied into the db for joins and such.

I was getting some mysterious C++ pointer exceptions on batched jobs which hinted at that. I'm still not sure if that was it, but it was fixed by re-adding the set_fds() facile dataset object. Was that getting stale somehow? I don't really get it still.