Re-exports main dplyr verbs.
I'm not sure what the best way to be a good citizen is, here. The package and
examples are almost always setting up the user to get down with dplyr mojo.
Should I just Depend on dplyr?

For now just re-exporting some of the main dplyr verbs. If the user wants more
they will have to `library(dplyr)`