Defensive coding in `assay_summary()` function.
If the `assay_info()` of an FDS type thing doesn't have "description" and "storage_mode" columns returned, we won't throw an error.

You may be wrapping external datastores in a facileapi and didn't 100% implement it. Not including such columns in the external source isn't a big deal, and there is no need to fail on that